Process PR comments
